+++ packages/Zooniverse-React-Components/src/index.js
@@ -7,3 +7,4 @@ export { default as ZooniverseLogotype } from './components/zooniverse-logotype'
 export { default as ZooniverseLogo } from './components/zooniverse-logo';
 export { default as DisplayNameSlugEditor } from './components/display-name-slug-editor';
 export { default as ImageSelector } from './components/image-selector';
+export { default as Paginator } from './components/paginator';