+++ packages/Zooniverse-React-Components/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.0] 2017-07-17
+### Added
+- Paginator component and tests
+
+### Fixed
+- Removed default syling from ZooniverseLogo
+- Added class method to DisplayNameSlugEditor to access state values
+
 ## [0.2.0] 2017-05-25
 ### Added
 - DisplayNameSlugEditor component and tests